backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
opencv-python
numpy
sqlalchemy
pydantic
langchain
langchain-openai
python-dotenv
easyocr
torch
torchvision

pytest
httpx
